% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{addTwo}
\alias{addTwo}
\title{Add two numbers}
\usage{
addTwo(x, y)
}
\arguments{
\item{x}{The first number to add}

\item{y}{The second number to add}
}
\value{
The sum of the 2 numbers
}
\description{
Given two numbers, add those two numbers together
}
\seealso{
\code{\link{hello}}
}
